# Optimize dictionary lookups