--- a/column_buffer_go18.go
+++ b/column_buffer_go18.go
@@ -10,21 +10,6 @@ import (
 	"github.com/segmentio/parquet-go/deprecated"
 )
 
-func makeArray[T any](s []T) array {
-	return array{
-		ptr: *(*unsafe.Pointer)(unsafe.Pointer(&s)),
-		len: len(s),
-	}
-}
-
-func makeSlice[T any](a array) []T {
-	return slice[T](a.ptr, a.len)
-}
-
-func slice[T any](p unsafe.Pointer, n int) []T {
-	return unsafe.Slice((*T)(p), n)
-}
-
 // writeRowsFunc is the type of functions that apply rows to a set of column
 // buffers.
 //