This PR adds optimizations for dictionary lookups when AVX-512 is available.

The following benchmarks compare between the pure Go and SIMD-optimized versions of the code, looking 1000 indexes from the dictionary at each iteration:
```
name                                        old time/op  new time/op  delta
Dictionary/Bounds/BOOLEAN                   29.3ns ± 0%  29.3ns ± 0%      ~     (p=0.329 n=10+10)
Dictionary/Bounds/INT32                      773ns ± 0%   238ns ± 0%   -69.22%  (p=0.000 n=9+10)
Dictionary/Bounds/INT64                      775ns ± 0%   317ns ± 1%   -59.12%  (p=0.000 n=10+9)
Dictionary/Bounds/INT96                     6.59µs ± 1%  6.28µs ± 1%    -4.76%  (p=0.000 n=10+10)
Dictionary/Bounds/FLOAT                     1.22µs ± 0%  0.24µs ± 0%   -80.39%  (p=0.000 n=8+10)
Dictionary/Bounds/DOUBLE                    1.22µs ± 0%  0.31µs ± 0%   -74.48%  (p=0.000 n=9+9)
Dictionary/Bounds/BYTE_ARRAY                10.2µs ± 0%   7.8µs ± 0%   -23.83%  (p=0.000 n=9+9)
Dictionary/Bounds/FIXED_LEN_BYTE_ARRAY(16)  8.33µs ± 0%  7.50µs ± 1%    -9.94%  (p=0.000 n=9+10)
Dictionary/Bounds/INT(32,false)              891ns ± 0%   237ns ± 0%   -73.44%  (p=0.000 n=10+10)
Dictionary/Bounds/INT(64,false)              892ns ± 0%   318ns ± 0%   -64.37%  (p=0.000 n=9+10)
Dictionary/Lookup/BOOLEAN                   2.32µs ± 0%  1.75µs ± 0%   -24.60%  (p=0.000 n=9+9)
Dictionary/Lookup/INT32                     1.40µs ± 0%  0.62µs ± 2%   -55.66%  (p=0.000 n=9+9)
Dictionary/Lookup/INT64                     1.60µs ± 0%  0.64µs ± 0%   -59.95%  (p=0.000 n=9+8)
Dictionary/Lookup/INT96                     20.4µs ± 2%  19.8µs ± 2%    -3.05%  (p=0.000 n=9+9)
Dictionary/Lookup/FLOAT                     1.53µs ± 0%  0.63µs ± 1%   -59.15%  (p=0.000 n=10+9)
Dictionary/Lookup/DOUBLE                    1.43µs ± 0%  0.65µs ± 4%   -54.87%  (p=0.000 n=8+10)
Dictionary/Lookup/BYTE_ARRAY                3.58µs ± 0%  0.98µs ± 1%   -72.51%  (p=0.000 n=8+9)
Dictionary/Lookup/FIXED_LEN_BYTE_ARRAY(16)  2.68µs ± 1%  0.91µs ± 0%   -65.95%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(32,false)             1.41µs ± 0%  0.61µs ± 1%   -56.85%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(64,false)             1.58µs ± 0%  0.73µs ± 0%   -53.53%  (p=0.000 n=8+10)

name                                        old value/s  new value/s  delta
Dictionary/Bounds/BOOLEAN                    34.1G ± 0%   34.1G ± 0%      ~     (p=0.353 n=10+10)
Dictionary/Bounds/INT32                      1.29G ± 0%   4.20G ± 0%  +224.88%  (p=0.000 n=9+10)
Dictionary/Bounds/INT64                      1.29G ± 0%   3.16G ± 1%  +144.60%  (p=0.000 n=10+9)
Dictionary/Bounds/INT96                       152M ± 1%    159M ± 1%    +5.00%  (p=0.000 n=10+10)
Dictionary/Bounds/FLOAT                       820M ± 0%   4181M ± 0%  +409.82%  (p=0.000 n=9+10)
Dictionary/Bounds/DOUBLE                      818M ± 0%   3205M ± 0%  +291.93%  (p=0.000 n=9+9)
Dictionary/Bounds/BYTE_ARRAY                 98.0M ± 0%  128.6M ± 0%   +31.29%  (p=0.000 n=9+9)
Dictionary/Bounds/FIXED_LEN_BYTE_ARRAY(16)    120M ± 0%    133M ± 1%   +11.03%  (p=0.000 n=9+10)
Dictionary/Bounds/INT(32,false)              1.12G ± 0%   4.22G ± 0%  +276.45%  (p=0.000 n=10+10)
Dictionary/Bounds/INT(64,false)              1.12G ± 0%   3.15G ± 0%  +180.65%  (p=0.000 n=9+10)
Dictionary/Lookup/BOOLEAN                     432M ± 0%    573M ± 0%   +32.63%  (p=0.000 n=9+9)
Dictionary/Lookup/INT32                       715M ± 0%   1613M ± 2%  +125.54%  (p=0.000 n=9+9)
Dictionary/Lookup/INT64                       627M ± 0%   1565M ± 0%  +149.70%  (p=0.000 n=9+8)
Dictionary/Lookup/INT96                      49.0M ± 2%   50.5M ± 2%    +3.15%  (p=0.000 n=9+9)
Dictionary/Lookup/FLOAT                       652M ± 0%   1596M ± 1%  +144.83%  (p=0.000 n=10+9)
Dictionary/Lookup/DOUBLE                      699M ± 0%   1549M ± 4%  +121.69%  (p=0.000 n=8+10)
Dictionary/Lookup/BYTE_ARRAY                  280M ± 0%   1017M ± 1%  +263.77%  (p=0.000 n=8+9)
Dictionary/Lookup/FIXED_LEN_BYTE_ARRAY(16)    373M ± 1%   1096M ± 0%  +193.64%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(32,false)               707M ± 0%   1639M ± 1%  +131.73%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(64,false)               632M ± 0%   1361M ± 0%  +115.32%  (p=0.000 n=9+10)
```