@achille-roussel 2022-06-07
For the sake of completeness, here's a snapshot of benchmarks comparing the current purego and amd64 versions of the code:

```
name                                        old time/op  new time/op  delta
Dictionary/Bounds/BOOLEAN                   29.3ns ± 0%  29.3ns ± 0%      ~     (p=0.481 n=10+10)
Dictionary/Bounds/INT32                      774ns ± 0%   238ns ± 0%   -69.25%  (p=0.000 n=10+10)
Dictionary/Bounds/INT64                      774ns ± 0%   317ns ± 1%   -59.06%  (p=0.000 n=10+9)
Dictionary/Bounds/INT96                     6.52µs ± 0%  6.28µs ± 1%    -3.70%  (p=0.000 n=10+10)
Dictionary/Bounds/FLOAT                     1.16µs ± 0%  0.24µs ± 0%   -79.37%  (p=0.000 n=9+10)
Dictionary/Bounds/DOUBLE                    1.16µs ± 0%  0.31µs ± 0%   -73.02%  (p=0.000 n=10+9)
Dictionary/Bounds/BYTE_ARRAY                9.72µs ± 0%  7.77µs ± 0%   -20.02%  (p=0.000 n=10+9)
Dictionary/Bounds/FIXED_LEN_BYTE_ARRAY(16)  8.53µs ± 0%  7.50µs ± 1%   -12.04%  (p=0.000 n=10+10)
Dictionary/Bounds/INT(32,false)              892ns ± 0%   237ns ± 0%   -73.46%  (p=0.000 n=9+10)
Dictionary/Bounds/INT(64,false)              892ns ± 0%   318ns ± 0%   -64.37%  (p=0.000 n=10+10)
Dictionary/Lookup/BOOLEAN                   2.30µs ± 0%  1.75µs ± 0%   -24.19%  (p=0.000 n=9+9)
Dictionary/Lookup/INT32                     1.43µs ± 1%  0.62µs ± 2%   -56.56%  (p=0.000 n=9+9)
Dictionary/Lookup/INT64                     1.40µs ± 0%  0.64µs ± 0%   -54.30%  (p=0.000 n=10+8)
Dictionary/Lookup/INT96                     20.4µs ± 2%  19.8µs ± 2%    -2.74%  (p=0.001 n=10+9)
Dictionary/Lookup/FLOAT                     1.63µs ± 0%  0.63µs ± 1%   -61.47%  (p=0.000 n=8+9)
Dictionary/Lookup/DOUBLE                    1.41µs ± 0%  0.65µs ± 4%   -54.34%  (p=0.000 n=9+10)
Dictionary/Lookup/BYTE_ARRAY                3.46µs ± 0%  0.98µs ± 1%   -71.58%  (p=0.000 n=9+9)
Dictionary/Lookup/FIXED_LEN_BYTE_ARRAY(16)  2.52µs ± 0%  0.91µs ± 0%   -63.78%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(32,false)             1.40µs ± 1%  0.61µs ± 1%   -56.47%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(64,false)             1.42µs ± 2%  0.73µs ± 0%   -48.42%  (p=0.000 n=9+10)

name                                        old value/s  new value/s  delta
Dictionary/Bounds/BOOLEAN                    34.1G ± 0%   34.1G ± 0%      ~     (p=0.739 n=10+10)
Dictionary/Bounds/INT32                      1.29G ± 0%   4.20G ± 0%  +225.23%  (p=0.000 n=10+10)
Dictionary/Bounds/INT64                      1.29G ± 0%   3.16G ± 1%  +144.27%  (p=0.000 n=10+9)
Dictionary/Bounds/INT96                       153M ± 0%    159M ± 1%    +3.84%  (p=0.000 n=10+10)
Dictionary/Bounds/FLOAT                       863M ± 0%   4181M ± 0%  +384.74%  (p=0.000 n=10+10)
Dictionary/Bounds/DOUBLE                      865M ± 0%   3205M ± 0%  +270.63%  (p=0.000 n=10+9)
Dictionary/Bounds/BYTE_ARRAY                  103M ± 0%    129M ± 0%   +25.03%  (p=0.000 n=10+9)
Dictionary/Bounds/FIXED_LEN_BYTE_ARRAY(16)    117M ± 0%    133M ± 1%   +13.69%  (p=0.000 n=10+10)
Dictionary/Bounds/INT(32,false)              1.12G ± 0%   4.22G ± 0%  +276.72%  (p=0.000 n=8+10)
Dictionary/Bounds/INT(64,false)              1.12G ± 0%   3.15G ± 0%  +180.65%  (p=0.000 n=10+10)
Dictionary/Lookup/BOOLEAN                     434M ± 0%    573M ± 0%   +31.89%  (p=0.000 n=9+9)
Dictionary/Lookup/INT32                       701M ± 1%   1613M ± 2%  +130.18%  (p=0.000 n=9+9)
Dictionary/Lookup/INT64                       715M ± 0%   1565M ± 0%  +118.81%  (p=0.000 n=10+8)
Dictionary/Lookup/INT96                      49.1M ± 3%   50.5M ± 2%    +2.81%  (p=0.001 n=10+9)
Dictionary/Lookup/FLOAT                       615M ± 0%   1596M ± 1%  +159.54%  (p=0.000 n=8+9)
Dictionary/Lookup/DOUBLE                      707M ± 0%   1549M ± 4%  +119.09%  (p=0.000 n=9+10)
Dictionary/Lookup/BYTE_ARRAY                  289M ± 0%   1017M ± 1%  +251.86%  (p=0.000 n=9+9)
Dictionary/Lookup/FIXED_LEN_BYTE_ARRAY(16)    397M ± 0%   1096M ± 0%  +176.05%  (p=0.000 n=9+9)
Dictionary/Lookup/INT(32,false)               713M ± 1%   1639M ± 1%  +129.75%  (p=0.000 n=10+9)
Dictionary/Lookup/INT(64,false)               702M ± 2%   1361M ± 0%   +93.85%  (p=0.000 n=9+10)
```